backend/requirements.txt
Django==5.1.1

psycopg2-binary==2.9.9
python-decouple==3.8
django-debug-toolbar==4.2.0
python-dateutil==2.9.0
gunicorn==20.1.0
redis==4.3.4
celery==5.2.7
requests==2.32.3
whitenoise==6.4.0
flower==2.0.1
pandas==2.2.2
pandas_ta==0.3.14b
numpy==1.26.4

plotly==5.23.0

django-prometheus
channels
djangorestframework
django-cors-headers
django-filter

# djangorestframework-simplejwt
# djangorestframework-api-key==2.*